requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pytz==2023.3
gunicorn==21.2.0
Werkzeug==2.3.7
pymongo==4.6.0
flask-pymongo==2.3.0 